fix error because the userId type is int

+++ src/main/java/com/yc/www/jfinal/service/article/ArticleService.java
@@ -31,7 +31,7 @@ public class ArticleService {
     //private static String rootPath = "/home/superuser/workspace/temp/ironman/articles"; //local
     private static String rootPath = "/apps/ironman/articles"; //production
     
-    public Article createArticle(String title, String content, int userId) {
+    public Article createArticle(String title, String content, long userId) {
         if(StringUtils.isBlank(title) || StringUtils.isBlank(content)) {
             return null;
         }
@@ -45,7 +45,7 @@ public class ArticleService {
         return SaveToDBUtil.saveModel(article);
     }
 
-    private String saveArticleToDisk(String title, String content, int userId) {
+    private String saveArticleToDisk(String title, String content, long userId) {
         String dirPath = rootPath + File.separator + userId + File.separator;
         File dir = new File(dirPath);
         if(!dir.exists()) {
@@ -99,7 +99,7 @@ public class ArticleService {
         return listAllArticles(-1);
     }
 
-    public List<ArticleVO> listAllArticles(int articleUserId) {
+    public List<ArticleVO> listAllArticles(long articleUserId) {
         StringBuilder sql = new StringBuilder("select * from article");
         if(articleUserId > 0) {
             sql.append(" where articleUserId = " + articleUserId);
